refactor(sui-ssr): refactor tests and code and remove rewire dep
