+++ packages/sui-ssr/package.json
@@ -43,6 +43,6 @@
     "ua-parser-js": "0.7.20"
   },
   "devDependencies": {
-    "@s-ui/test": "4.11.0-beta.1"
+    "@s-ui/test": "4"
   }
 }
