+++ packages/sui-ssr/server/utils/index.js
@@ -1,0 +1,24 @@
+import path from 'path'
+import fs from 'fs'
+import config from '../config'
+import utilsFactory from './factory'
+
+const {
+  isMultiSite,
+  hostFromReq,
+  publicFolder,
+  useStaticsByHost,
+  readHtmlTemplate,
+  hrTimeToMs,
+  buildRequestUrl
+} = utilsFactory({path, fs, config})
+
+export {
+  isMultiSite,
+  hostFromReq,
+  publicFolder,
+  useStaticsByHost,
+  readHtmlTemplate,
+  hrTimeToMs,
+  buildRequestUrl
+}
