+++ packages/sui-ssr/test/server/fixtures/index.js
@@ -5,9 +5,25 @@ export const getMockedRequest = hostname => ({
   protocol: 'http'
 })
 
-export const multiSiteMapping = {
-  'www.bikes.com': 'bikes',
-  'www.trucks.com': 'trucks',
-  'dev.trucks.com': 'trucks',
-  default: 'bikes'
+export const ssrConfig = {
+  criticalCss: {
+    protocol: 'https',
+    host: 'www.bikes.com'
+  }
+}
+
+export const ssrMultiSiteConfig = {
+  criticalCss: {
+    protocol: 'https',
+    host: {
+      bikes: 'www.bikes.com',
+      trucks: 'www.trucks.com'
+    }
+  },
+  multiSite: {
+    'www.bikes.com': 'bikes',
+    'www.trucks.com': 'trucks',
+    'dev.trucks.com': 'trucks',
+    default: 'bikes'
+  }
 }
