+++ packages/sui-ssr/test/server/utilsSpec.js
@@ -1,27 +1,15 @@
 import {expect} from 'chai'
-import {publicFolder, __RewireAPI__ as utilsRewireAPI} from '../../server/utils'
-import {getMockedRequest, multiSiteMapping} from './fixtures'
+import {publicFolder} from '../../server/utils'
+import {getMockedRequest} from './fixtures'
+import {publicFolderWithMultiSiteConfig} from './fixtures/utils'
 
 describe('[sui-ssr] Utils', () => {
   describe('Public folder', () => {
     describe('In a multi site project', () => {
-      beforeEach(() => {
-        utilsRewireAPI.__Rewire__('multiSiteMapping', multiSiteMapping)
-        utilsRewireAPI.__Rewire__(
-          'multiSiteKeys',
-          Object.keys(multiSiteMapping)
-        )
-        utilsRewireAPI.__Rewire__('isMultiSite', true)
-      })
-
-      afterEach(() => {
-        utilsRewireAPI.__ResetDependency__('multiSiteMapping')
-        utilsRewireAPI.__ResetDependency__('multiSiteKeys')
-        utilsRewireAPI.__ResetDependency__('isMultiSite')
-      })
-
       it('Should build the public folder name properly', () => {
-        const folderName = publicFolder(getMockedRequest('www.bikes.com'))
+        const folderName = publicFolderWithMultiSiteConfig(
+          getMockedRequest('www.bikes.com')
+        )
         expect(folderName).to.equal('public-bikes')
       })
     })
